+++ sshd-core/src/main/java/org/apache/sshd/agent/unix/UnixAgentFactory.java
@@ -30,9 +30,8 @@ import org.apache.sshd.agent.SshAgentFactory;
 import org.apache.sshd.agent.SshAgentServer;
 import org.apache.sshd.common.Factory;
 import org.apache.sshd.common.FactoryManager;
-import org.apache.sshd.common.NamedFactory;
 import org.apache.sshd.common.SshException;
-import org.apache.sshd.common.channel.Channel;
+import org.apache.sshd.common.channel.ChannelFactory;
 import org.apache.sshd.common.session.ConnectionService;
 import org.apache.sshd.common.session.Session;
 import org.apache.sshd.common.util.GenericUtils;
@@ -44,9 +43,9 @@ import org.apache.sshd.server.session.ServerSession;
  * @author <a href="mailto:dev@mina.apache.org">Apache MINA SSHD Project</a>
  */
 public class UnixAgentFactory implements SshAgentFactory {
-    public static final List<NamedFactory<Channel>> DEFAULT_FORWARDING_CHANNELS =
+    public static final List<ChannelFactory> DEFAULT_FORWARDING_CHANNELS =
         Collections.unmodifiableList(
-            Arrays.<NamedFactory<Channel>>asList(
+            Arrays.asList(
                 ChannelAgentForwardingFactory.OPENSSH,
                 ChannelAgentForwardingFactory.IETF));
 
@@ -65,11 +64,11 @@ public class UnixAgentFactory implements SshAgentFactory {
     }
 
     @Override
-    public List<NamedFactory<Channel>> getChannelForwardingFactories(FactoryManager manager) {
+    public List<ChannelFactory> getChannelForwardingFactories(FactoryManager manager) {
         if (executorServiceFactory != null) {
             return DEFAULT_FORWARDING_CHANNELS.stream()
-                    .map(cf -> new ChannelAgentForwardingFactory(cf.getName(), executorServiceFactory))
-                    .collect(Collectors.toList());
+                .map(cf -> new ChannelAgentForwardingFactory(cf.getName(), executorServiceFactory))
+                .collect(Collectors.toList());
         } else {
             return DEFAULT_FORWARDING_CHANNELS;
         }