+++ sshd-core/src/main/java/org/apache/sshd/common/channel/ChannelFactory.java
@@ -19,13 +19,38 @@
 
 package org.apache.sshd.common.channel;
 
-import org.apache.sshd.common.NamedFactory;
+import java.io.IOException;
+import java.util.Collection;
+
+import org.apache.sshd.common.NamedResource;
+import org.apache.sshd.common.session.Session;
 
 /**
  * @author <a href="mailto:dev@mina.apache.org">Apache MINA SSHD Project</a>
  */
-// CHECKSTYLE:OFF
-public interface ChannelFactory extends NamedFactory<Channel> {
-    // nothing extra
+public interface ChannelFactory extends NamedResource {
+    /**
+     * @param session The {@link Session} through which the request is made
+     * @return The relevant {@link Channel}
+     * @throws IOException If failed to create the requested instance
+     */
+    Channel createChannel(Session session) throws IOException;
+
+    /**
+     * @param session The {@link Session} through which the request is made
+     * @param factories The available factories
+     * @param name The required factory name to use
+     * @return The created {@link Channel} - {@code null} if no match found
+     * @throws IOException If failed to create the requested instance
+     */
+    static Channel createChannel(
+            Session session, Collection<? extends ChannelFactory> factories, String name)
+                throws IOException {
+        ChannelFactory f = NamedResource.findByName(name, String.CASE_INSENSITIVE_ORDER, factories);
+        if (f != null) {
+            return f.createChannel(session);
+        } else {
+            return null;
+        }
+    }
 }
-//CHECKSTYLE:ON