+++ sshd-core/src/test/java/org/apache/sshd/common/channel/WindowTest.java
@@ -109,7 +109,7 @@ public class WindowTest extends BaseTestSupport {
         sshd.setChannelFactories(Arrays.asList(
                 new ChannelSessionFactory() {
                     @Override
-                    public Channel create() {
+                    public Channel createChannel(Session session) throws IOException {
                         return new ChannelSession() {
                             @SuppressWarnings("synthetic-access")
                             @Override