+++ CHANGES.md
@@ -25,6 +25,9 @@ has been made
 * `UserAuthFactory` is a proper interface and it has been refactored to contain a
 `createUserAuth` method that accepts the session instance through which the request is made.
 
+* `ChannelFactory` is a proper interface and it has been refactored to contain a
+`createChannel` method that accepts the session instance through which the request is made.
+
 ## Minor code helpers
 
 * `SessionListener` supports `sessionPeerIdentificationReceived` that is invoked once successful
@@ -39,4 +42,4 @@ for the server's identification before sending its own.
 
 * [SSHD-934](https://issues.apache.org/jira/browse/SSHD-934) - Fixed ECDSA public key encoding into OpenSSH format.
 
-* [SSHD-937](https://issues.apache.org/jira/browse/SSHD-937) - Provide session instance when creating a subsystem or user authentication.
+* [SSHD-937](https://issues.apache.org/jira/browse/SSHD-937) - Provide session instance when creating a subsystem, user authentication, channel.